moved write_nz_bbox_to_file() and get_catchment_area() from tide_query_location.py to main_tide_slr.py
